os/arch/arm/src: Move mpu-reg.h file out of chip specific code
Since mpu-reg.h does not contain any chip specific code, we will
move it out of imxrt folder to the armv7-m folder.
